refactor
- split corpus modules out into files
